fix: Remove DefaultResolvers by moving them to the companion object
